task_4

+++ src/main/cpp/t02_frac_2.cpp
@@ -18,7 +18,7 @@
 using namespace std;
 
 int t02_frac_2() {
-    float n=0;
+    double n=0;
     int x=0;
 
     cin>>n;
